requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Mail==0.10.0
Flask-WTF==1.2.1
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
tornado==6.4.1
urllib3==2.2.3
Werkzeug==3.0.3
WTForms==3.1.2
